% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_resources.R
\name{addExternalResources}
\alias{addExternalResources}
\title{Add Shiny static resources to Shiny Server}
\usage{
addExternalResources()
}
\value{
adds external resources to the Shiny app by returning a
\code{tags$head} element that includes:
\itemize{
\item Resource Paths
\item JavaScript and CSS scripts
\item Favicon
\item Font Links
\item Custom JS Handles
@keywords internal
}
}
\description{
Make scripts available to their components.
}
