% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mining.R
\name{ageDist}
\alias{ageDist}
\title{Distribution of platforms by age groups}
\usage{
ageDist(long)
}
\arguments{
\item{ultra_combo}{all the cheerleader data with ages}
}
\value{
side effect is to return ggplot2 object with platform count
metrics by age groups.
}
\description{
Distribution of platforms by age groups
}
\keyword{internal}
