% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mining.R
\name{ageJitterDist}
\alias{ageJitterDist}
\title{Create interactive age distribution}
\usage{
ageJitterDist(long, team_data)
}
\arguments{
\item{ultra_combo}{All the cheerleader and team data}
}
\value{
plotly object
}
\description{
Create interactive age distribution
}
\keyword{internal}
