% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Server function for the application}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{Internal parameter for \code{{shiny}}}

\item{output}{Internal parameter for \code{{shiny}}}

\item{session}{Internal parameter for \code{{shiny}}}
}
\description{
Server function for the application
}
