% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mining.R
\name{backup}
\alias{backup}
\title{Backup Data}
\usage{
backup()
}
\value{
side effect is to make the historical backup data that will later
be extracted for 'historical' trendlines within the leaderboards.
}
\description{
Backup /data and /www and all contents. Store in timestamped folder.
}
\details{
ex. KBO_Cheerleaders_Backup_20240817_201241/data
/www/cheerleader_img
/www/social_icons
/www/team_cap
/www/team_img
/www/team_logo
}
\keyword{internal}
