% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mining.R
\name{cheerData}
\alias{cheerData}
\title{Get Cheerleader Data}
\usage{
cheerData(wiki_url, team_cheerleaders, values)
}
\arguments{
\item{wiki_url}{namuwiki cheerleader page prefix}

\item{team_cheerleaders}{data.frame with team, cheerleader and namu wiki url}

\item{values}{values to match on that exist in all cheerleader tables.}
}
\value{
list all cheerleaders bio(HTML, data.frame) and social media links
}
\description{
Iterate over team cheerleader URLs to make the list of HTML, data.frame
and link vectors.
}
\keyword{internal}
