% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mining.R
\name{cheerTableIndex}
\alias{cheerTableIndex}
\title{Get the index of the cheerleader table}
\usage{
cheerTableIndex(tables, keyword)
}
\arguments{
\item{tables}{all the tables on the team page}

\item{keyword}{keyword vector to identify common string which all
cheerleader tables share}
}
\value{
integer of the table index the cheerleader list belongs to.
}
\description{
Get the index of the cheerleader table
}
\keyword{internal}
