% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mining.R
\name{extractAge}
\alias{extractAge}
\title{Extracts age from the "Birth" or "Age" row of the cheerleaders table.}
\usage{
extractAge(text)
}
\arguments{
\item{text}{the value (X2) record of matching "Birthday" (X1) record.}
}
\value{
Cheerleaders Age or NA if could not extract.
}
\description{
Extracts age from the "Birth" or "Age" row of the cheerleaders table.
}
