% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mining.R
\name{extractChannelID}
\alias{extractChannelID}
\title{Extract YouTube channel ID from a given URL.}
\usage{
extractChannelID(url)
}
\arguments{
\item{url}{Cheerleader YouTube URL}
}
\value{
channel_id
}
\description{
The YouTube channel ID will either be found in the url or in the raw HTML
for the page. The pattern to search of is a 24 character length string.
}
\keyword{internal}
