% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mining.R
\name{fatDistroPlot}
\alias{fatDistroPlot}
\title{Followers Across Teams Distribution Graphs}
\usage{
fatDistroPlot(ultra_combo)
}
\arguments{
\item{ultra_combo}{all the datas}
}
\value{
list of distribution graphs
\itemize{
\item normal distribution
\item capped outliers @ 100k
\item log scale 1k - 1M
\item 95\% percentile
}
}
\description{
Followers Across Teams Distribution Graphs
}
\keyword{internal}
