% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mining.R
\name{fatPlot}
\alias{fatPlot}
\title{Aggregate Team Followers plot}
\usage{
fatPlot(ultra_combo)
}
\arguments{
\item{ultra_combo}{followers across teams data.frame}
}
\value{
ggplot2 with team logo images
}
\description{
Allows faster loading times.
}
\keyword{internal}
