% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mining.R
\name{followersAcrossTeams}
\alias{followersAcrossTeams}
\title{Aggregate Team Followers plot}
\usage{
followersAcrossTeams(fat)
}
\arguments{
\item{fat}{followers across teams data.frame}
}
\value{
ggplot2 with team logo images
}
\description{
Allows faster loading times.
}
