% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mining.R
\name{getCheerleaderPage}
\alias{getCheerleaderPage}
\title{Get Cheerleader Page}
\usage{
getCheerleaderPage(wiki_url, cheerleader_url, values)
}
\arguments{
\item{wiki_url}{string wiki url prefix}

\item{cheerleader_url}{string wiki cheerleader url suffix}

\item{values}{values that exist in the cheerleader table to match upon}
}
\value{
returns a named list of
\itemize{
\item name/index Cheerleaders name
\item data.frame the cheerleaders biography
\item list the cheerleaders biography as html node
\item character vector of social media links
}
}
\description{
Will iterate over all the cheerleaders page links and extract their biography.
}
\details{
Cheerleader Name
}
\keyword{internal}
