% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mining.R
\name{getCheerleaderPhotos}
\alias{getCheerleaderPhotos}
\title{Get Cheerleader Photo}
\usage{
getCheerleaderPhotos(bio_tables, cheer_data)
}
\arguments{
\item{cheer_data}{all cheerleader table data}

\item{bio_table}{html node from cheer_data$cheerleader$bio_table}
}
\value{
side effect is to save the image to /cheerleader_img
}
\description{
Get Cheerleader Photo
}
