% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mining.R
\name{getCheerleaders}
\alias{getCheerleaders}
\title{Get cheerleaders}
\usage{
getCheerleaders(team_url)
}
\arguments{
\item{team_url}{team_data$url}
}
\value{
character vector of cheerleader url links and names
}
\description{
Query the team wiki page for a list of cheerleaders and url to their
personal pages.
}
