% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mining.R
\name{getInstagram}
\alias{getInstagram}
\title{Get Cheerleader Instagram Statistics.}
\usage{
getInstagram(cheer_data)
}
\arguments{
\item{cheer_data}{data.frame of cheerleader data.}
}
\value{
data.frame with Instagram statistics
\itemize{
\item Cheerleader name
\item Instagram account name
\item Instagram followers
\item Cheerleader team
\item Type of social media platform
}
}
\description{
Get Cheerleader Instagram Statistics.
}
\keyword{internal}
