% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mining.R
\name{getTeamCap}
\alias{getTeamCap}
\title{Save team cap insignia}
\usage{
getTeamCap(team_caps)
}
\arguments{
\item{team_caps}{team_caps}
}
\value{
side effect is to download the team_cap/* images.
}
\description{
Save team cap insignia
}
