% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mining.R
\name{getTeamCheerleaders}
\alias{getTeamCheerleaders}
\title{Get Team Page Cheerleaders}
\usage{
getTeamCheerleaders(team_url)
}
\arguments{
\item{team_url}{team_data$url}
}
\value{
data.frame with the team, cheerleader name and their page url.
}
\description{
Query the team wiki page for a list of cheerleaders and url to their
personal pages. Iterates through a list of all the cheerleaders for
all teams.
}
\keyword{internal}
