% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mining.R
\name{getTeamLogos}
\alias{getTeamLogos}
\title{Save team logos}
\usage{
getTeamLogos(team_logos)
}
\arguments{
\item{team_logos}{team_logos}
}
\value{
side effect is to download the team_logos/* images.
}
\description{
Download the team logo.
}
\keyword{internal}
