% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mining.R
\name{getTeamPhotos}
\alias{getTeamPhotos}
\title{Get Team Photos}
\usage{
getTeamPhotos(team_data)
}
\arguments{
\item{team_data}{}
}
\value{
side effect is to download the team photo for each team and store
it in the local directory.
}
\description{
Iterate through all the team page data to download each team photo.
}
