% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mining.R
\name{getTikTok}
\alias{getTikTok}
\title{Get Cheerleader TikTok Statistics.}
\usage{
getTikTok(cheer_data)
}
\arguments{
\item{cheer_data}{data.frame of cheerleader data.}
}
\value{
data.frame with Instagram statistics
\itemize{
\item Cheerleader name
\item TikTok account name
\item TikTok follower count
\item TikTok likes count
\item Cheerleader team
\item Type of social media platform
}
}
\description{
Get Cheerleader TikTok Statistics.
}
