% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mining.R
\name{getYouTube}
\alias{getYouTube}
\title{Get Cheerleader YouTube Statistics.}
\usage{
getYouTube(cheer_data)
}
\arguments{
\item{cheer_data}{data.frame of cheerleader data.}
}
\value{
data.frame with YouTube statistics
\itemize{
\item Cheerleader name
\item YouTube channel
\item Subscriber count
\item Video view count
\item Video count
\item Cheerleader team
\item Type of social media platform
}
}
\description{
Get Cheerleader YouTube Statistics.
}
