% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mining.R
\name{load_historic_data}
\alias{load_historic_data}
\title{Create historical ultra_combo}
\usage{
load_historic_data(base_path = "../")
}
\arguments{
\item{base_path}{backup folders directory}
}
\value{
data.frame with cheerleader historical social media metrics
}
\description{
Create historical ultra_combo
}
