% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mining.R
\name{loadHistoricalData}
\alias{loadHistoricalData}
\title{Create historical ultra_combo}
\usage{
loadHistoricalData(base_path = "../")
}
\arguments{
\item{base_path}{backup folders directory}
}
\value{
data.frame with cheerleader historical social media metrics
}
\description{
Will be run weekly after backups to gather historical social media metrics.
Used in leaderboard sparkline plots.
}
\keyword{internal}
