% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{makeCheerleader}
\alias{makeCheerleader}
\title{Generate Cheerleader Image, Biography and Social Metric cards.}
\usage{
makeCheerleader(td, smm, cheerleader, cheerPhoto, cheerBio)
}
\arguments{
\item{td}{team_data reactive}

\item{smm}{social media metrics reactive}

\item{cheerleader}{input$cheerleader selected cheerleader}

\item{cheerPhoto}{namespace for cheerPhoto ui}

\item{cheerBio}{namespace for cheerBio ui}
}
\value{
Cheerleader bslib UI elements
}
\description{
Generate Cheerleader Image, Biography and Social Metric cards.
}
