% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mining.R
\name{makeFat}
\alias{makeFat}
\title{Aggregate YouTube Instagram and TikTok follower data}
\usage{
makeFat(youtube, instagram, tiktok)
}
\arguments{
\item{youtube}{cheerleader YouTube data}

\item{instagram}{cheerleader Instagram data}

\item{tiktok}{cheerleader TikTok data}
}
\value{
data.frame: team name, logo, color and followers per cheerleader
}
\description{
Aggregate YouTube Instagram and TikTok follower data
}
