% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{makegtInst}
\alias{makegtInst}
\title{Generate \code{gt} tables for Instagram Leaderboard}
\usage{
makegtInst(ultra_combo)
}
\arguments{
\item{ultra_combo}{aggregate cheerleader social media metrics.}
}
\value{
gt
}
\description{
Function takes all cheerleader social media metrics and generates
the Instagram leaderboard for top 5 cheerleaders with the most
Instagram followers.
}
