% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{makeLeaderboards}
\alias{makeLeaderboards}
\title{Generate Leaderboard Cards}
\usage{
makeLeaderboards(leaderYT, leaderInst, leaderTT)
}
\arguments{
\item{leaderYT}{namespace for youtube leaderboard}

\item{leaderInst}{namespace for instagram leaderboard}

\item{leaderTT}{namespace for tiktok leaderboard}
}
\value{
nice bslib card layout
}
\description{
Houses the 3 gt Leaderboard tables
}
\keyword{internal}
