% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{makeReactable}
\alias{makeReactable}
\title{Generate the reactable table}
\usage{
makeReactable(leader_data)
}
\arguments{
\item{leader_data}{reactive list of ultra_combo and teams}
}
\value{
\code{reactable::reactable}
}
\description{
Reactable table with Cheerleader Teams and individual cheerleader
stats such as total followers, YouTube subscribers, Instagram followers,
and TikTok followers arranged in descending order.
}
\keyword{internal}
