% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{makeStatsPage}
\alias{makeStatsPage}
\title{Generate stats tab}
\usage{
makeStatsPage()
}
\value{
bslib::page_fillable
}
\description{
This is the main plot(s) viewable when the app loads.
ggplot with aggregate team followers counts and distributions
of followers by platform.
}
\details{
These are all the UI cards that contain the actual plot outputs
}
