% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{makeTeam}
\alias{makeTeam}
\title{Make the team photo, logo and hat insignia card layouts.}
\usage{
makeTeam(td, teamPhoto, teamLogo, capInsignia)
}
\arguments{
\item{td}{reactive team data}

\item{teamPhoto}{namespace for team photo ui}

\item{teamLogo}{namespace for team logo ui}

\item{capInsignia}{namespace for cap insignia ui}
}
\value{
nice bslib card layout
}
\description{
Make the team photo, logo and hat insignia card layouts.
}
\keyword{internal}
