% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mining.R
\name{makeUltraLong}
\alias{makeUltraLong}
\title{Ultra Combo transposed long for aggregate followers}
\usage{
makeUltraLong(ultra_combo)
}
\arguments{
\item{ultra_combo}{all the cheerleader data}
}
\value{
reshape long ultra_combo
}
\description{
Ultra Combo transposed long for aggregate followers
}
