% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modCheer.R
\name{mod_cheer_server}
\alias{mod_cheer_server}
\title{Title}
\usage{
mod_cheer_server(id, td, smm)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the
module's UI function.}

\item{td}{reactive team_data}

\item{smm}{reactive social media metrics}
}
\value{
Side effect is to dynamically generate and update UI components
for selecting and displaying cheerleader information:
\itemize{
\item Radio buttons for selecting a cheerleader
\item UI Elements to display the cheerleaders photo
\item DT:: for displaying the cheerleaders biography or data.
@keywords internal
}
}
\description{
Title
}
