% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modCheer.R
\name{mod_cheer_ui}
\alias{mod_cheer_ui}
\title{Cheerleader UI}
\usage{
mod_cheer_ui(id)
}
\arguments{
\item{id}{A unique identifier string for the module’s UI.}
}
\value{
list of Shiny reactive outputs.
}
\description{
Cheerleader UI
}
\keyword{internal}
