% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modLeader.R
\name{mod_leaderboard_server}
\alias{mod_leaderboard_server}
\title{Leaderboard Server}
\usage{
mod_leaderboard_server(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the
module's UI function.}
}
\value{
Side effect is to render and update the three leaderboards:
\itemize{
\item leaderYT
\item leaderInst
\item leaderTT
using gt tables and dynamically generate UI elements for display.
}
}
\description{
Generate the leaderboard UI that holds the 3 cheerleader leaderboard
cards for
\itemize{
\item YouTube
\item Instagram
\item TikTok
}
}
\keyword{internal}
