% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modLeader.R
\name{mod_leaderboard_ui}
\alias{mod_leaderboard_ui}
\title{Leaderboard UI}
\usage{
mod_leaderboard_ui(id)
}
\arguments{
\item{id}{A unique identifier string for the module’s UI.}
}
\value{
Shiny reactive UI output
}
\description{
Leaderboard UI
}
