% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modReact.R
\name{mod_react_server}
\alias{mod_react_server}
\title{Reactable Server}
\usage{
mod_react_server(id, td)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the
module's UI function.}

\item{td}{reactive team_data}
}
\value{
side effect render the reactable
}
\description{
A reactable with expanding rows, for each team display all the cheerleaders
for that team, and their followers statistics.
}
