% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modReact.R
\name{mod_react_ui}
\alias{mod_react_ui}
\title{Reactable UI}
\usage{
mod_react_ui(id)
}
\arguments{
\item{id}{A unique identifier string for the module’s UI.}
}
\value{
Shiny reactive output
}
\description{
Reactable UI
}
\keyword{internal}
