% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modSong.R
\name{mod_song_ui}
\alias{mod_song_ui}
\title{Song UI}
\usage{
mod_song_ui(id)
}
\arguments{
\item{id}{A unique identifier string for the module’s UI.}
}
\value{
list of Shiny reactive input and output:
\itemize{
\item Team song toggle switch
\item UI for team song
\item UI for piki piki song
}
}
\description{
Song UI
}
