% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modStats.R
\name{mod_stats_server}
\alias{mod_stats_server}
\title{Team followers server}
\usage{
mod_stats_server(id, agg_follow)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{agg_follow}{reactive which sums/sorts total followers by
\itemize{
\item team
\item color
\item image
}}
}
\value{
ggplot2 with interactive team logo
}
\description{
Team followers server
}
