% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modStats.R
\name{mod_stats_ui}
\alias{mod_stats_ui}
\title{User Interface module for team followers}
\usage{
mod_stats_ui(id)
}
\arguments{
\item{id}{@param id A unique identifier string for the module’s UI.}
}
\value{
shiny::uiOutput
}
\description{
User Interface module for team followers
}
