% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modTeam.R
\name{mod_team_server}
\alias{mod_team_server}
\title{Team Server}
\usage{
mod_team_server(id, td)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the
module's UI function.}

\item{td}{reactive team_data object}
}
\value{
side effect is to render ui which consists of
\itemize{
\item team photo
\item team logo
\item team cap insignia
}
}
\description{
Team Server
}
