% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modTeam.R
\name{mod_team_ui}
\alias{mod_team_ui}
\title{Team UI}
\usage{
mod_team_ui(id)
}
\arguments{
\item{id}{@param id A unique identifier string for the module’s UI.}
}
\value{
list of shiny reactive input and output.
}
\description{
Team UI
}
