% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{team_data}
\alias{team_data}
\title{Team Data}
\format{
A data frame with X rows and Y columns:
\describe{
\item{name}{A character vector of team names.}
\item{url}{A character vector of team namuwiki urls.}
\item{color}{A character vector of team themesong urls on YouTube.}
}
}
\source{
"Extracted from namuwiki pages."
}
\usage{
team_data
}
\description{
A dataset containing information about various cheerleading teams.
}
\details{
This dataset includes columns such as team names, locations, and performance metrics.
}
\examples{
data(team_data)
head(team_data)
}
\keyword{datasets}
