% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{updateUI}
\alias{updateUI}
\title{Manage visibility and updates for team/cheerleader UI's}
\usage{
updateUI(session, state, team = NULL, cheerleader = NULL)
}
\arguments{
\item{session}{Current Shiny Session Environment}

\item{state}{Current state of reactive input to handle.}

\item{selected_team}{input$'team-team'}

\item{selected_cheerleader}{input$'cheer-cheerleader'}
}
\value{
commands to be sent to update the app UI.
}
\description{
Reduce redundancy improve readability
}
