% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_resources.R
\name{addExternalResources}
\alias{addExternalResources}
\title{Add Shiny static resources to Shiny Server}
\usage{
addExternalResources()
}
\description{
Make scripts available to their components.
}
