% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{makegtYT}
\alias{makegtYT}
\title{Generate \code{gt} tables for YouTube Leaderboard}
\usage{
makegtYT(historic)
}
\arguments{
\item{ultra_combo}{aggregate cheerleader social media metrics.}
}
\value{
gt
}
\description{
Function takes all cheerleader social media metrics and generates
the YouTube leaderboard for top (x) cheerleaders with the most
YouTube followers.
}
\keyword{internal}
