% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mining.R
\name{ultraCombo}
\alias{ultraCombo}
\title{All cheerleader/team data with social media metrics}
\usage{
ultraCombo(team_cheerleaders, cheer_data, youtube, instagram, tiktok)
}
\arguments{
\item{team_cheerleaders}{Cheerleader team, name and wiki url}

\item{cheer_data}{cheerleader data table/links}

\item{youtube}{youtube}

\item{instagram}{instagram}

\item{tiktok}{tiktok}
}
\value{
data.frame with 14 columns
}
\description{
All cheerleader/team data with social media metrics
}
\keyword{internal}
